requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
distlib==0.3.7
Django==5.1.6
django-cors-headers==4.7.0
django-csp==3.7
django-phonenumber-field==7.1.0
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.13.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
loguru==0.7.2
packaging==24.2
pbr==6.0.0
phonenumbers==8.13.14
platformdirs==3.11.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pygame==2.5.2
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dotenv==1.0.1
pytz==2023.3.post1
ratelimit==2.2.1
requests==2.32.3
service-identity==24.2.0
setuptools==68.2.2
sqlparse==0.5.3
stevedore==5.1.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2022.1
urllib3==2.3.0
virtualenv==20.24.6
virtualenv-clone==0.5.7
virtualenvwrapper==6.0.0
virtualenvwrapper-win==1.2.7
win32-setctime==1.1.0
zope.interface==7.2
